+++ lampo-common/Cargo.toml
@@ -17,10 +17,6 @@ bitcoin = { version = "0.30.2", features = ["serde"] }
 clightningrpc-conf = { git = "https://github.com/laanwj/cln4rust.git", branch = "master" }
 crossbeam-channel = "0.5.8"
 anyhow = "1.0.70"
-triggered = "0.1.2"
-tokio = { version = "1", features = ["full"] }
-ctrlc = { version = "3.1", features = ["termination"] }
-tonic = "0.9"
 colored = "1.9"
 log = { version = "0.4", features = ["std"] }
 chrono = { version = "0.4", features = ["std"], default-features = false }
