+++ lampo-vls/Cargo.toml
@@ -10,6 +10,7 @@ lampo-common = { path = "../lampo-common" }
 triggered = "0.1.2"
 url = "2.5.0"
 async-trait = "0.1"
+tonic = "0.9"
 tokio = { version = "1", features = ["full"] }
 ctrlc = { version = "3.1", features = ["termination"] }
