+++ lampo-vls/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod protocol_handler;
 pub mod signer;
-pub mod frontend_starter;
+pub mod util;
 
 use std::net::{Ipv4Addr, SocketAddr};
 use std::str::FromStr;
@@ -8,6 +8,7 @@ use std::sync::{Arc, Mutex};
 
 use protocol_handler::GrpcProtocolHandler;
 use signer::VLSSignerPort;
+use util::{AsyncRuntime, VLSKeysManager, Shutter};
 use vls_proxy::grpc::adapter::HsmdService;
 use vls_proxy::grpc::incoming::TcpIncoming;
 use vls_proxy::grpc::signer_loop::InitMessageCache;
@@ -20,14 +21,11 @@ use url::Url;
 
 use lampo_common::bitcoin::Network;
 use lampo_common::conf::LampoConf;
-use lampo_common::keys::{GrpcKeysManager, AsyncRuntime, KeysManagerFactory, Shutter};
 
-pub struct VLSKeysManagerFactory;
+pub struct VLSKeys;
 
-impl KeysManagerFactory for VLSKeysManagerFactory {
-    type GenericKeysManager = KeysManagerClient;
-
-    fn create_keys_manager(&self, conf: Arc<LampoConf>, seed: &[u8; 32], vls_port: Option<u16>, shutter: Option<Shutter>) -> GrpcKeysManager {
+impl VLSKeys {
+    pub fn create_keys_manager(&self, conf: Arc<LampoConf>, seed: &[u8; 32], vls_port: Option<u16>, shutter: Option<Shutter>) -> VLSKeysManager {
         let config = SignerConfig::new(conf, *seed, vls_port, shutter);
         SignerType::GrpcRemote.create_keys_manager(config)
     }
@@ -65,33 +63,36 @@ pub enum SignerType {
 
 impl SignerType {
     /// Method to create a signer based on the signer type
-    fn create_keys_manager(&self, config: SignerConfig) -> GrpcKeysManager {
+    fn create_keys_manager(&self, config: SignerConfig) -> VLSKeysManager {
         match self {
-            SignerType::InProcess => unimplemented!(),
-            // SignerType::InProcess => {
-            //     // This will create a handler that will manage the VLS protocol operations
-            //     let protocol_handler = Arc::new(protocol_handler::InProcessProtocolHandler::new(config.network, &config.seed));
+            SignerType::InProcess => {
+                let async_runtime = Arc::new(AsyncRuntime::new());
 
-            //     let signer_port = Arc::new(signer::VLSSignerPort::new(protocol_handler.clone()));
-            //     // This factory manages data sources but doesn't actually do anything (dummy).
-            //     let source_factory = Arc::new(SourceFactory::new(config.lampo_data_dir, config.network));
-            //     // The SignerPortFront provide a client RPC interface to the core MultiSigner and Node objects via a communications link.
-            //     let signer_port_front = Arc::new(SignerPortFront::new(signer_port, config.network));
-            //     // The frontend acts like a proxy to handle communication between the Signer and the Node
-            //     let frontend = Frontend::new(
-            //         signer_port_front,
-            //         source_factory,
-            //         config.bitcoin_rpc_url,
-            //         config.shutdown_signal,
-            //     );
+                let (keys_manager, server_handler) = async_runtime.block_on(async {
+                    let protocol_handler = Arc::new(protocol_handler::InProcessProtocolHandler::new(config.network, &config.seed));
 
-            //     FrontendStarter::start_frontend(Arc::new(frontend));
-            //     // Starts the frontend (probably discuss is this the right place to start or not)
-            //     // frontend.start();
-            //     // Similar to KeysManager from LDK but here as all Key related operations are happening on the
-            //     // signer, thus we need a client to facilitate that
-            //     KeysManagerClient::new(protocol_handler, config.network.to_string())
-            // }
+                    let signer_port = Arc::new(signer::VLSSignerPort::new(protocol_handler.clone()));
+                    // This factory manages data sources but doesn't actually do anything (dummy).
+                    let source_factory = Arc::new(SourceFactory::new(config.lampo_data_dir, config.network));
+                    // The SignerPortFront provide a client RPC interface to the core MultiSigner and Node objects via a communications link.
+                    let signer_port_front = Arc::new(SignerPortFront::new(signer_port, config.network));
+                    // The frontend acts like a proxy to handle communication between the Signer and the Node
+                    let frontend = Frontend::new(
+                        signer_port_front,
+                        source_factory,
+                        config.bitcoin_rpc_url,
+                        config.shutdown_signal,
+                    );
+
+                    frontend.start();
+                    
+                    (KeysManagerClient::new(protocol_handler, config.network.to_string()), None)
+                });
+
+                VLSKeysManager::new(async_runtime, keys_manager, server_handler)
+
+            },
+
             SignerType::GrpcRemote => {
     
                 let async_runtime = Arc::new(AsyncRuntime::new());
@@ -108,14 +109,7 @@ impl SignerType {
 
                     let hello = signer_handle.spawn(server.start(incoming, shutter.signal.clone()));
 
-                    // let transport = Arc::new(
-                    //     signer_handle
-                    //         .spawn(GrpcProtocolHandler::new(sender))
-                    //         .await
-                    //         .expect("join")
-                    //         .expect("gRPC transport init"),
-                    // );
-                    let transport = Arc::new(GrpcProtocolHandler::new(sender, async_runtime.clone()).await.expect("FUCKED!!!!"));
+                    let transport = Arc::new(GrpcProtocolHandler::new(sender, async_runtime.clone()).await.expect("Cannot create gRPC transport"));
 
                     let source_factory = Arc::new(SourceFactory::new(config.lampo_data_dir, config.network));
                     let signer_port = Arc::new(VLSSignerPort::new(transport.clone()));
@@ -130,7 +124,7 @@ impl SignerType {
                     (KeysManagerClient::new(transport, config.network.to_string()), hello)
                 });
 
-                GrpcKeysManager::new(async_runtime, keys_manager, server_handle)
+                VLSKeysManager::new(async_runtime, keys_manager, Option::from(server_handle))
             }
         }
     }
